# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid request body
        type: string
    type: object
  handlers.EventResponse:
    properties:
      completedAt:
        type: string
      date:
        type: string
      groupOrdering:
        $ref: '#/definitions/models.GroupOrdering'
      hasWinnersGroup:
        type: boolean
      isComplete:
        type: boolean
      isFinals:
        type: boolean
      machines:
        items:
          $ref: '#/definitions/models.Machine'
        type: array
      name:
        type: string
      players:
        items:
          $ref: '#/definitions/models.Player'
        type: array
      season:
        $ref: '#/definitions/models.Season'
      seasonID:
        type: integer
      seedingMethod:
        $ref: '#/definitions/models.SeedingMethod'
    type: object
  handlers.LeagueResponse:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      dateCreated:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      location:
        example: 123 Main St
        type: string
      name:
        example: Downtown Pinball League
        type: string
      owner:
        $ref: '#/definitions/models.SwaggerUser'
      ownerID:
        example: 1
        type: integer
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  handlers.ListResponse:
    properties:
      data: {}
    type: object
  handlers.PlayerResponse:
    properties:
      ifpaNumber:
        type: string
      league:
        $ref: '#/definitions/models.League'
      leagueID:
        type: integer
      name:
        type: string
    type: object
...
  models.GroupOrdering:
    enum:
    - RANDOM
    - SEEDED
    type: string
    x-enum-varnames:
    - GroupOrderingRandom
    - GroupOrderingSeeded
  models.League:
    properties:
      dateCreated:
        type: string
      location:
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      ownerID:
        type: integer
    type: object
  models.Machine:
    properties:
      created_at:
        description: LastUpdatedAt time.Time `json:"last_updated_at"`
        type: string
      ipdb_id:
        type: integer
      is_alias:
        type: boolean
      is_group:
        type: boolean
      is_pinball:
        type: boolean
      name:
        type: string
      opdb_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      year:
        description: Manufacturer  string    `json:"manufacturer"`
        type: integer
    type: object
  models.Player:
    properties:
      ifpaNumber:
        type: string
      league:
        $ref: '#/definitions/models.League'
      leagueID:
        type: integer
      name:
        type: string
    type: object
  models.PointDistributionMap:
    additionalProperties:
      items:
        type: number
      type: array
    type: object
  models.Season:
    properties:
      countingGames:
        type: integer
      created_at:
        type: string
      dateCreated:
        type: string
      eventCount:
        type: integer
      hasFinals:
        type: boolean
      league:
        $ref: '#/definitions/models.League'
      leagueID:
        type: integer
      name:
        type: string
      pointDistribution:
        $ref: '#/definitions/models.PointDistributionMap'
    type: object
  models.SeedingMethod:
    enum:
    - AVERAGE
    - RANK
    - RANDOM
    - IFPA_RANK
    type: string
    x-enum-varnames:
    - SeedingMethodAverage
    - SeedingMethodRank
    - SeedingMethodRandom
    - SeedingMethodIFPARank
...
  models.SwaggerUser:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      firstName:
        example: John
        type: string
      id:
        example: 1
        type: integer
      lastName:
        example: Doe
        type: string
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  models.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      leagues:
        items:
          $ref: '#/definitions/models.League'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Pinball League Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pinball League API
  version: "1.0"
paths:
...
  /leagues:
    get:
      description: Get a list of all pinball leagues
      produces:
      - application/json
      responses:
        "200":
          description: List of leagues
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.LeagueResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all leagues
      tags:
      - leagues
  /leagues/{leagueID}:
    get:
      description: Get detailed information about a specific league
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: League details
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.LeagueResponse'
              type: object
        "400":
          description: Invalid league ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get league by ID
      tags:
      - leagues
  /leagues/{leagueID}/players:
    get:
      description: Get a list of all players in a specific league
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of players
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.PlayerResponse'
                  type: array
              type: object
        "400":
          description: Invalid league ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List players in a league
      tags:
      - leagues
  /leagues/{leagueID}/players/ifpa:
    post:
      consumes:
      - application/json
      description: Add multiple players to a league using their IFPA numbers
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      - description: IFPA numbers
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Players added successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.PlayerResponse'
                  type: array
              type: object
        "400":
          description: Invalid request body or league ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Add players to league by IFPA numbers
      tags:
      - leagues
...
  /leagues/{leagueID}/seasons/{seasonID}/events:
    get:
      description: Get a list of all events for a specific season
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      - description: Season ID
        in: path
        name: seasonID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of events
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.EventResponse'
                  type: array
              type: object
        "400":
          description: Invalid league ID or season ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List events for a season
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event for a specific season
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      - description: Season ID
        in: path
        name: seasonID
        required: true
        type: string
      - description: Event details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EventResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new event
      tags:
      - events
  /leagues/{leagueID}/seasons/{seasonID}/events/{eventID}:
    get:
      description: Get detailed information about a specific event
      parameters:
      - description: League ID
        in: path
        name: leagueID
        required: true
        type: string
      - description: Season ID
        in: path
        name: seasonID
        required: true
        type: string
      - description: Event ID
        in: path
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event details
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.EventResponse'
              type: object
        "400":
          description: Invalid league ID, season ID, or event ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get event by ID
      tags:
      - events
  /leagues/create:
    post:
      consumes:
      - application/json
      description: Create a new pinball league
      parameters:
      - description: League details
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: League created successfully
          schema:
            allOf:
            - $ref: '#/definitions/handlers.ListResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.LeagueResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new league
      tags:
      - leagues
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"